chore: general project cleanup and background_detector implementation
