+++ lib/utils/dimensions.dart
@@ -8,6 +8,7 @@ const double spacingXXL = 24.0;
 const double spacingXXXL = 28.0;
 const double spacingXXXXL = 32.0;
 const double spacingXXXXXL = 36.0;
+const double spacingXXXXXXL = 42.0;
 
 const fontSizeXS = 8.0;
 const fontSizeS = 12.0;
@@ -17,3 +18,6 @@ const fontSizeXL = 24.0;
 const fontSizeXXL = 32.0;
 
 const iconSize = 72.0;
+
+const logoHeight = 58.0;
+const logoWidth = 222.0;
