--- lib/view/screen/call_screen.dart
@@ -1,70 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:logger/logger.dart';
-import 'package:provider/provider.dart';
-import 'package:telnyx_flutter_webrtc/view/telnyx_client_view_model.dart';
-import 'package:telnyx_flutter_webrtc/view/widgets/dialpad_widget.dart';
-import 'package:telnyx_webrtc/call.dart';
-
-class CallScreen extends StatefulWidget {
-  const CallScreen({super.key, this.call});
-
-  final Call? call;
-
-  @override
-  State<CallScreen> createState() => _CallScreenState();
-}
-
-class _CallScreenState extends State<CallScreen> {
-  final logger = Logger();
-  TextEditingController callInputController = TextEditingController();
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Ongoing Call'),
-      ),
-      body: SingleChildScrollView(
-        child: Center(
-          child: Column(
-            children: [
-              const SizedBox(height: 16),
-              Text(widget.call?.sessionDestinationNumber ?? 'Unknown Caller'),
-              const SizedBox(height: 8),
-              const SizedBox(height: 8),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: [
-                  IconButton(
-                    onPressed: () {
-                      print('mic');
-                      Provider.of<TelnyxClientViewModel>(context, listen: false)
-                          .muteUnmute();
-                    },
-                    icon: const Icon(Icons.mic),
-                  ),
-                  IconButton(
-                    onPressed: () {
-                      print('speakerphone');
-                      Provider.of<TelnyxClientViewModel>(context, listen: false)
-                          .toggleSpeakerPhone();
-                    },
-                    icon: const Icon(Icons.volume_up),
-                  ),
-                  IconButton(
-                    onPressed: () {
-                      print('pause');
-                      Provider.of<TelnyxClientViewModel>(context, listen: false)
-                          .holdUnhold();
-                    },
-                    icon: const Icon(Icons.pause),
-                  ),
-                ],
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
