+++ lib/view/widgets/call_controls/call_controls.dart
@@ -5,6 +5,9 @@ import 'package:telnyx_flutter_webrtc/utils/asset_paths.dart';
 import 'package:telnyx_flutter_webrtc/utils/dimensions.dart';
 import 'package:telnyx_flutter_webrtc/utils/theme.dart';
 import 'package:telnyx_flutter_webrtc/view/telnyx_client_view_model.dart';
+import 'package:telnyx_flutter_webrtc/view/widgets/call_controls/buttons/call_buttons.dart';
+import 'package:telnyx_flutter_webrtc/view/widgets/call_controls/call_invitation.dart';
+import 'package:telnyx_flutter_webrtc/view/widgets/call_controls/ongoing_call_controls.dart';
 import 'package:telnyx_flutter_webrtc/view/widgets/dialpad_widget.dart';
 
 class CallControls extends StatefulWidget {
@@ -91,186 +94,4 @@ class _CallControlsState extends State<CallControls> {
       ],
     );
   }
-}
-
-class OnGoingCallControls extends StatelessWidget {
-  const OnGoingCallControls({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: [
-            CallControlButton(
-              enabledIcon: Icons.mic,
-              disabledIcon: Icons.mic_off,
-              isDisabled: context.select<TelnyxClientViewModel, bool>(
-                (txClient) => txClient.muteState,
-              ),
-              onToggle: () {
-                context.read<TelnyxClientViewModel>().muteUnmute();
-              },
-            ),
-            DeclineButton(
-              onPressed: () {
-                context.read<TelnyxClientViewModel>().endCall();
-              },
-            ),
-            CallControlButton(
-              enabledIcon: Icons.volume_up,
-              disabledIcon: Icons.volume_off,
-              isDisabled: context.select<TelnyxClientViewModel, bool>(
-                (txClient) => txClient.speakerPhoneState,
-              ),
-              onToggle: () {
-                context.read<TelnyxClientViewModel>().toggleSpeakerPhone();
-              },
-            ),
-          ],
-        ),
-        SizedBox(height: spacingM),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: [
-            CallControlButton(
-              enabledIcon: Icons.play_arrow,
-              disabledIcon: Icons.pause,
-              isDisabled: context.select<TelnyxClientViewModel, bool>(
-                (txClient) => txClient.holdState,
-              ),
-              onToggle: () {
-                context.read<TelnyxClientViewModel>().holdUnhold();
-              },
-            ),
-            SizedBox(width: iconSize),
-            CallControlButton(
-              enabledIcon: Icons.dialpad,
-              disabledIcon: Icons.dialpad,
-              isDisabled: false,
-              onToggle: () {
-                showModalBottomSheet(
-                  context: context,
-                  backgroundColor: Colors.transparent,
-                  isScrollControlled: true,
-                  builder: (context) {
-                    return Padding(
-                      padding: EdgeInsets.only(
-                        bottom: MediaQuery.of(context).viewInsets.bottom,
-                      ),
-                      child: DialPad(
-                        onDigitPressed: (digit) {
-                          context.read<TelnyxClientViewModel>().dtmf(digit);
-                        },
-                      ),
-                    );
-                  },
-                );
-              },
-            ),
-          ],
-        ),
-      ],
-    );
-  }
-}
-
-class CallControlButton extends StatefulWidget {
-  final IconData enabledIcon;
-  final IconData disabledIcon;
-  final bool isDisabled;
-  final VoidCallback onToggle;
-
-  const CallControlButton({
-    super.key,
-    required this.enabledIcon,
-    required this.disabledIcon,
-    required this.isDisabled,
-    required this.onToggle,
-  });
-
-  @override
-  State<CallControlButton> createState() => _CallControlButtonState();
-}
-
-class _CallControlButtonState extends State<CallControlButton> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: iconSize,
-      height: iconSize,
-      decoration: BoxDecoration(
-        color: call_control_color,
-        shape: BoxShape.circle,
-      ),
-      child: IconButton(
-        icon:
-            Icon(widget.isDisabled ? widget.disabledIcon : widget.enabledIcon),
-        onPressed: widget.onToggle,
-      ),
-    );
-  }
-}
-
-class CallInvitation extends StatelessWidget {
-  final VoidCallback onAccept;
-  final VoidCallback onDecline;
-
-  const CallInvitation({
-    super.key,
-    required this.onAccept,
-    required this.onDecline,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-      children: [
-        CallButton(
-          onPressed: onAccept,
-        ),
-        SizedBox(width: spacingM),
-        DeclineButton(
-          onPressed: onDecline,
-        ),
-      ],
-    );
-  }
-}
-
-abstract class BaseButton extends StatelessWidget {
-  final VoidCallback onPressed;
-  final String iconPath;
-
-  const BaseButton({
-    super.key,
-    required this.onPressed,
-    required this.iconPath,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
-      onTap: onPressed,
-      child: SvgPicture.asset(
-        iconPath,
-        width: iconSize,
-        height: iconSize,
-      ),
-    );
-  }
-}
-
-class CallButton extends BaseButton {
-  const CallButton({super.key, required VoidCallback onPressed})
-      : super(onPressed: onPressed, iconPath: green_call_icon);
-}
-
-class DeclineButton extends BaseButton {
-  const DeclineButton({super.key, required VoidCallback onPressed})
-      : super(onPressed: onPressed, iconPath: red_decline_icon);
 }
